% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-id.R
\name{assign_id}
\alias{assign_id}
\title{Assign ID values to distinct dataframe rows}
\usage{
assign_id(
  df,
  vars,
  id_name,
  prefix = NULL,
  seq_start = 1,
  digits = NULL,
  promote = TRUE
)
}
\arguments{
\item{df}{A dataframe.}

\item{vars}{Variables for determining row uniqueness.}

\item{id_name}{The name for the ID variable.}

\item{prefix}{An optional prefix added to all ID values.}

\item{seq_start}{The start number of the ID sequence.}

\item{digits}{The number of digits in the ID, including leading zeros. This
is used when larger than the minimum required.}

\item{promote}{Logical. If \code{TRUE}, the ID variable is the first column in the
output.}
}
\value{
A dataframe.
}
\description{
Assign sequential ID values to distinct dataframe rows based on the variables
specified in \code{vars}.
}
\details{
Row names are not preserved.
}
\examples{
df <- data.frame(
  var1 = rep(LETTERS[1:3], 4),
  var2 = 1:2
)

assign_id(
  df,
  id_name = "id",
  prefix = "id",
  vars = colnames(df)
)

}
