% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-date-order.R
\name{check_date_order}
\alias{check_date_order}
\title{Check date order}
\usage{
check_date_order(...)
}
\arguments{
\item{...}{Date vectors entered in their expected order, earliest to latest.}
}
\value{
A logical vector the same length as the inputs.
}
\description{
Check if dates are ordered from earliest to latest.
}
\details{
Any number of date vectors can be checked. Elements at the same index are
compared. \code{NA}s are ignored.
}
\examples{
df <- data.frame(
  d1 = seq.Date(
    as.Date("1980-01-01"),
    as.Date("1980-12-31"),
    by = "month"
  ),
  d2 = seq.Date(
    as.Date("1990-01-01"),
    as.Date("1990-12-31"),
    by = "month"
  ),
  d3 = seq.Date(
    as.Date("1985-01-01"),
    as.Date("1996-01-01"),
    by = "year"
  )
)

df$ordered <- check_date_order(df$d1, df$d2, df$d3)

df

}
