% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-to-css-rgba.R
\name{color_to_css_rgba}
\alias{color_to_css_rgba}
\title{Color to CSS rgba() function call}
\usage{
color_to_css_rgba(color, alpha = 1)
}
\arguments{
\item{color}{A vector of color names and/or hex color values.}

\item{alpha}{A vector of alpha channel values from 0 to 1.}
}
\value{
A character vector of the same length as \code{color}.
}
\description{
Convert color names or hex values to CSS rgba() function calls.
}
\examples{
color_to_css_rgba(c("red", "#FF0000"), alpha = c(.2, .4))

}
