% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-to-hex.R
\name{color_to_hex}
\alias{color_to_hex}
\title{Color to hexadecimal value}
\usage{
color_to_hex(color, alpha = NULL)
}
\arguments{
\item{color}{A vector of color names and/or hex strings.}

\item{alpha}{A vector of alpha channel values from 0 to 1.}
}
\value{
A character vector of the same length as \code{color}.
}
\description{
Convert color names to their hexadecimal values.
}
\details{
When \code{alpha} is null, a hex value is returned without an alpha channel. Hex
values with or without the alpha channel are accepted as input. However, the
\code{alpha} argument determines if the alpha channel is included in the output.
}
\examples{
color_to_hex(c("blue", "white"))

color_to_hex(c("lightgreen", "#90EE90"), alpha = c(.2, .4))

}
