% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast-color.R
\name{contrast_color}
\alias{contrast_color}
\title{Get a contrasting color}
\usage{
contrast_color(color)
}
\arguments{
\item{color}{A vector of color names and/or hex color values.}
}
\value{
A character vector of the same length as \code{color}.
}
\description{
Based on the luminance value of \code{color}, get the color, either black or
white, that is most contrasting.
}
\examples{
contrast_color(c("black", "white", "#FF0000", "#888"))

}
