% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-doc.R
\name{convert_doc}
\alias{convert_doc}
\title{Convert a document to another format using Pandoc}
\usage{
convert_doc(file1, file2)
}
\arguments{
\item{file1, file2}{The names of the file to convert and the destination file.
Include both filename extensions (this is how the "from" and "to" file
formats are determined).}
}
\value{
A character vector if \code{file2} is a markdown or typst file, otherwise \code{NULL}.
}
\description{
Convert a document from one format to another by calling the \code{pandoc} system
command.
}
\details{
Pandoc must be installed. See https://pandoc.org/.
}
\examples{
\dontrun{
# Convert from Word to markdown
convert_doc("my-doc.docx", "my-doc.md")
}

}
