% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-dupes.R
\name{find_dupes}
\alias{find_dupes}
\title{Find duplicated rows}
\usage{
find_dupes(df, vars)
}
\arguments{
\item{df}{A dataframe.}

\item{vars}{Variables whose values determine which rows are identified as
duplicates.}
}
\value{
A dataframe consisting of only duplicated rows, or \code{NULL} if no
duplicates are found
}
\description{
Find the duplicated rows in a dataframe.
}
\details{
The output is sorted with duplicate rows adjacent to each other. A \code{dupe_id}
variable is added which numbers each duplicate set.
}
\examples{
df <- data.frame(
  row_id = paste0("r", 1:12),
  x = c(101:104, 101, 112:114, 101:102, 123:124),
  y = rep(LETTERS[1:4], 3)
)

find_dupes(df, vars = c("x", "y"))

}
