% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-sorted.R
\name{is_sorted}
\alias{is_sorted}
\title{Check if values are sorted}
\usage{
is_sorted(...)
}
\arguments{
\item{...}{One vector, or multiple vectors of the same length.}
}
\value{
If there is one input, a logical value. If there are multiple inputs,
a logical vector the same length as the inputs.
}
\description{
Check if values are sorted in increasing order from left to right.
}
\details{
Any number of vectors can be compared. If the input is a single vector, the
order of elements in that vector is checked. However, if the input is
multiple vectors, the order of elements across all vectors at each index is
checked. \code{NA}s are ignored.
}
\examples{
is_sorted(letters)

is_sorted(rev(letters))

x <- 1:5
y <- c(2, 3, NA, 0, 5)
is_sorted(x, y)

df <- data.frame(
  d1 = seq.Date(
    as.Date("1980-01-01"),
    as.Date("1980-12-31"),
    by = "month"
  ),
  d2 = seq.Date(
    as.Date("1990-01-01"),
    as.Date("1990-12-31"),
    by = "month"
  ),
  d3 = seq.Date(
    as.Date("1985-01-01"),
    as.Date("1996-01-01"),
    by = "year"
  )
)

df$sorted <- is_sorted(df$d1, df$d2, df$d3)

df

}
