% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable-table.R
\name{readable_table}
\alias{readable_table}
\title{Make a readable table}
\usage{
readable_table(df, width)
}
\arguments{
\item{df}{A dataframe.}

\item{width}{The maximum width of each column.}
}
\value{
A dataframe.
}
\description{
Transform a dataframe so that it can be written to a text file as an easily
readable table. Row names are not preserved.
}
\examples{
df <- data.frame(
  x = c(
    "One line of text   ",
    "This text should be on 2 lines"
  ),
  y = c(
    "This text should be on 4 lines etc etc etc etc etc etc",
    "   One line of text"
  )
)

df2 <- readable_table(df, width = 16)

\dontrun{
write.table(
  df2,
  file = "readable-table-example.txt",
  sep = "\t",
  quote = FALSE,
  row.names = FALSE
)
}

}
