% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round-ties-away.R
\name{round_ties_away}
\alias{round_ties_away}
\title{Round to nearest value, ties away from 0}
\usage{
round_ties_away(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{An integer indicating the number of decimal (when positive) or
whole number (when zero or negative) places to round \code{x} to. The rounding
place is equivalent to \eqn{1 / 10^{digits}}. \code{1} rounds to the nearest
tenths place, \code{0} rounds to the nearest ones place, and \code{-1} rounds to the
nearest tens place. (This behaves the same as the \code{digits} argument in
\code{\link[base:round]{base::round()}}.)}
}
\value{
A numeric vector the same length as \code{x}.
}
\description{
Round numbers to the nearest whole or decimal value with ties (numbers
halfway between two values, i.e., ending in 5) rounding to the nearest value
in the direction away from zero.
}
\details{
\code{\link[base:round]{base::round()}} rounds numbers using the "ties to even" method, resulting in
both \code{round(1.5)} and \code{round(2.5)} evaluating to \code{2}. This method attempts to
counteract bias in data sets that include numbers that have already been
rounded using the "ties away from 0" method (see links below for more
detail). Alternatively, \code{round_ties_away()} rounds numbers using the "ties
away from 0" method, resulting in \code{round_ties_away(1.5)} evaluating to \code{2}
and \code{round_ties_away(2.5)} evaluating to \code{3}. This may be preferred in some
scenarios, e.g., when presenting results.

Code in this function was copied from \code{\link[janitor:round_half_up]{janitor::round_half_up()}} (version
2.2.0), which was itself copied from a Stack Overflow forum (see the link
below).

Resources:
\itemize{
\item \href{https://en.wikipedia.org/wiki/Rounding}{Wikipedia entry on Rounding}
\item \href{https://stackoverflow.com/questions/12688717/round-up-from-5}{Stack Overflow discussion and early source of the function code}
}
}
\examples{
x1 <- seq(-2.5, 2.5, by = 0.5)
x2 <- seq(-0.025, 0.025, by = 0.005)
x3 <- seq(-250, 250, by = 50)

x1
round(x1)
round_ties_away(x1)

x2
round(x2, digits = 2)
round_ties_away(x2, digits = 2)

x3
round(x3, digits = -2)
round_ties_away(x3, digits = -2)

}
