% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project-structure.R
\name{setup_project_structure}
\alias{setup_project_structure}
\title{Setup project structure}
\usage{
setup_project_structure(
  root = getwd(),
  dirs = c("data/1-source", "data/2-aux", "data/3-final", "output", "scripts")
)
}
\arguments{
\item{root}{The project root directory (defaults to the current working
directory).}

\item{dirs}{The subdirectories to create, as a character vector. The defaults
are described in the Details section.}
}
\value{
\code{root} (invisibly).
}
\description{
Create subdirectories for an R project.
}
\details{
By default, these subdirectories are created under \code{root}:
\itemize{
\item \verb{data/1-source/}: Unmodified source data originating outside of the project.
\item \verb{data/2-aux/}: Intermediate data created within the project and/or anything not considered to be source or final data.
\item \verb{data/3-final/}: Final data products in native R formats only.
\item \verb{output/}: Final data products for consumption outside R (CSV, XLSX, XML, HTML, PDF, etc.).
\item \verb{scripts/}: All script files.
}

Directories are created recursively by \code{\link[base:files2]{base::dir.create()}}.
}
\examples{
\dontrun{
setup_project_structure()
}

}
