% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-rmd-template.R
\name{use_rmd_template}
\alias{use_rmd_template}
\title{Use an RMD template from this package}
\usage{
use_rmd_template(filename, template)
}
\arguments{
\item{filename}{The file name for the script.}

\item{template}{The name of the RMD template.}
}
\value{
\code{filename} (invisibly).
}
\description{
This is a wrapper for \code{\link[rmarkdown:draft]{rmarkdown::draft()}}. It copies the contents of the
specified template folder into the \verb{scripts/} directory (created by
\code{\link[=setup_project_structure]{setup_project_structure()}}).
}
\examples{
\dontrun{
use_rmd_template(filename = "my-analysis.rmd", template = "nb1")
}

}
