% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-id.R
\name{assign_id}
\alias{assign_id}
\title{Assign ID values to distinct dataframe rows}
\usage{
assign_id(
  df,
  id_name,
  prefix = NULL,
  vars = colnames(df),
  seq_start = 1,
  digits = NULL,
  promote = TRUE
)
}
\arguments{
\item{df}{A dataframe.}

\item{id_name}{A name for the new ID variable.}

\item{prefix}{An optional prefix of characters to add to all ID values.}

\item{vars}{The variable names whose unique values the new ID will be based
on. Defaults to \code{colnames(df)}.}

\item{seq_start}{An integer for the beginning of the ID sequence. Defaults to
\code{1}.}

\item{digits}{An integer for the number of digits in the ID (including
leading zeros). Defaults to \code{nchar(nrow(df) + seq_start)}.}

\item{promote}{A logical value. If \code{TRUE}, move the ID column to the first
position in \code{df}.}
}
\value{
A dataframe.
}
\description{
Assign sequential ID values to distinct dataframe rows based on the variables
specified in \code{vars}.
}
\examples{
df <- data.frame(
  row_id = paste("r", 1:24),
  a = rep(c("a", "b", "c"), times = 8),
  b = c(1, 1, 2, 2),
  c = c(TRUE, FALSE)
)

df

assign_id(
  df,
  id_name = "id",
  prefix = "id",
  vars = c("a", "b", "c")
)

}
