% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-date-order.R
\name{check_date_order}
\alias{check_date_order}
\title{Check date order}
\usage{
check_date_order(...)
}
\arguments{
\item{...}{Date vectors entered in their expected order, earliest to latest.}
}
\value{
A logical vector the same length as the inputs.
}
\description{
Check if dates are ordered from earliest to latest.
}
\details{
Any number of date vectors can be checked. Elements at the same index are
compared. \code{NA}s are ignored.
}
\examples{
dates <- seq.Date(
  as.Date("2023-01-01", "\%Y-\%m-\%d"),
  as.Date("2023-12-31", "\%Y-\%m-\%d"),
  by = "day"
)
dates <- c(dates, rep(NA, times = 10))

n_rows <- 500

df <- data.frame(
  d1 = sample(dates, n_rows, replace = TRUE),
  d2 = sample(dates, n_rows, replace = TRUE),
  d3 = sample(dates, n_rows, replace = TRUE),
  d4 = sample(dates, n_rows, replace = TRUE)
)

df$ordered <- check_date_order(df$d1, df$d2, df$d3, df$d4)

}
