% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-dupes.R
\name{find_dupes}
\alias{find_dupes}
\title{Find duplicated rows}
\usage{
find_dupes(df, vars)
}
\arguments{
\item{df}{A dataframe.}

\item{vars}{Variables whose values determine which rows are identified as
duplicates.}
}
\value{
A dataframe consisting of only duplicated rows.
}
\description{
Keep only the duplicated rows in a dataframe.
}
\details{
The output is sorted with duplicate rows adjacent to each other. A \code{dupe_id}
variable is added which numbers each duplicate set.
}
\examples{
df <- data.frame(
  row_id = paste0("r", 1:12),
  var1 = rep(LETTERS[1:4], 3),
  var2 = c(1:4, 1, 12:14, 1:2, 23:24)
)

find_dupes(df, vars = c("var1", "var2"))

}
