% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-colnames.R
\name{fix_colnames}
\alias{fix_colnames}
\title{Fix column names}
\usage{
fix_colnames(x)
}
\arguments{
\item{x}{A vector of column names.}
}
\value{
A character vector the same length as \code{x}.
}
\description{
Change column names to snake case and remove or replace non-word characters.
}
\examples{
cols <- c(
  "Messy/Unpleasant Column Names",
  "checkThisCamelCase",
  "Letters674next0to827654digits4",
  "ordinal numbers: 1st 32nd 43rd 54th 900th",
  "this&THAT",
  "THIS/that",
  " Hello, World!!! ",
  "hyphenated-words",
  "words in 'quotes' or even \"quotes\""
)

fix_colnames(cols)

}
