% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-colnames.R
\name{fix_colnames}
\alias{fix_colnames}
\title{Fix column names}
\usage{
fix_colnames(x)
}
\arguments{
\item{x}{A vector of column names.}
}
\value{
A character vector the same length as \code{x}.
}
\description{
Change column names to snake case.
}
\examples{
s <- paste(
  "Some words and stuff plus punctuation",
  ". , : ; ? ! / * @ # - _ \" ' [ ] { } ( )",
  "And symbols",
  "| ` = + ^ ~ < > $",
  "And these",
  "\% &",
  "checkThisCamelCase",
  "Letters674next0to827654digits4",
  "THIS/THAT"
)

fix_colnames(s)

}
