% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.R
\name{pct}
\alias{pct}
\title{Calculate percentage}
\usage{
pct(n, total, digits = 1)
}
\arguments{
\item{n, total}{Numeric vectors, or objects that can be coerced to numeric
vectors. If the lengths are unequal, the shorter vector must have length of 1
for it to be recycled.}

\item{digits}{An integer indicating the number of decimal (when positive) or
whole number (when zero or negative) places to round the output. (See the
same argument in \code{\link[=round_ties_away]{round_ties_away()}}.)}
}
\value{
A numeric vector the same length as the longer of \code{n} or \code{total.}
}
\description{
Calculate a percentage using the formula \eqn{n * 100 / total}. Rounding is
done by \code{\link[=round_ties_away]{round_ties_away()}}.
}
\examples{
x <- sample(200:300, size = 10, replace = TRUE)

pct(x, 300, digits = 2)

}
