% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project.R
\name{setup_project}
\alias{setup_project}
\title{Setup project structure}
\usage{
setup_project(
  root = getwd(),
  type = c("analysis", "dashboard"),
  dirs = NULL,
  rmd = TRUE
)
}
\arguments{
\item{root}{The project root directory (defaults to the current working
directory).}

\item{type}{The type of project to create (\code{"analysis"} or \code{"dashboard"}).}

\item{dirs}{The subdirectories to create, as a character vector. The defaults
are described in the Details section.}

\item{rmd}{Call \code{\link[=use_rmd_template]{use_rmd_template()}} if TRUE (with defaults only, currently).}
}
\value{
\code{root} (invisibly).
}
\description{
Create subdirectories for an R project.
}
\details{
\subsection{Default subdirectories created under \code{root}}{

For \code{type = "analysis"}:
\itemize{
\item \verb{data/}
\itemize{
\item \verb{1-source/}: Unmodified source data files originating outside of the
project.
\item \verb{2-final/}: Final data products (native R formats only).
}
\item \verb{output/}: Outputs for consumption outside R (e.g., CSV, XLSX, XML, HTML,
PDF).
\item \verb{scripts/}: All script files (e.g., R, RMD, QMD).
}

For \code{type = "dashboard"}:
\itemize{
\item \verb{data/}: All data files.
\item \verb{scripts/}: All script files (e.g., R, RMD, QMD).
}

Directories are created recursively by \code{\link[base:files2]{base::dir.create()}}.
}
}
\examples{
\dontrun{
setup_project(type = "analysis")
}

}
