% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project.R
\name{setup_project}
\alias{setup_project}
\title{Setup project structure}
\usage{
setup_project(
  root = getwd(),
  dirs = c("data/1-source", "data/2-final", "output", "scripts"),
  rmd = TRUE
)
}
\arguments{
\item{root}{The project root directory (defaults to the current working
directory).}

\item{dirs}{The subdirectories to create, as a character vector. The defaults
are described in the Details section.}

\item{rmd}{Call \code{\link[=use_rmd_template]{use_rmd_template()}} if TRUE (with defaults only, currently).}
}
\value{
\code{root} (invisibly).
}
\description{
Create subdirectories for an R project.
}
\details{
By default, these subdirectories are created under \code{root}:
\itemize{
\item \verb{data/1-source/}: Unmodified source data originating outside of the project.
\item \verb{data/2-final/}: Final data products in native R formats only.
\item \verb{output/}: Outputs for consumption outside R (CSV, XLSX, XML, HTML, PDF, etc.).
\item \verb{scripts/}: All script files.
}

Directories are created recursively by \code{\link[base:files2]{base::dir.create()}}.
}
\examples{
\dontrun{
setup_project()
}

}
