% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project-structure.R
\name{setup_project_structure}
\alias{setup_project_structure}
\title{Setup project structure}
\usage{
setup_project_structure(dir = getwd())
}
\arguments{
\item{dir}{The project root directory. Defaults to the current working
directory.}
}
\value{
\code{dir} (invisibly).
}
\description{
This function creates subdirectories for a data project.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{project-root-dir
|--- data
     |--- 1-source
     |--- 2-aux
     |--- 3-final
|--- output
|--- scripts
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
setup_project_structure()
}

}
