% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project-structure.R
\name{setup_project_structure}
\alias{setup_project_structure}
\title{Setup project structure}
\usage{
setup_project_structure(dir = getwd())
}
\arguments{
\item{dir}{The project root directory (defaults to the current working
directory).}
}
\value{
\code{dir} (invisibly).
}
\description{
Create subdirectories for a data analysis project.
}
\details{
This function creates the following subdirectories under the hypothetical
project root directory \verb{my-proj/}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{my-proj/
|--- data/
     |--- 1-source/
     |--- 2-aux/
     |--- 3-final/
|--- output/
|--- scripts/
}\if{html}{\out{</div>}}\tabular{ll}{
   Folder \tab Intended content \cr
   \verb{data/1-source/} \tab Unmodified source data originating outside of the project. \cr
   \verb{data/2-aux/} \tab Intermediate data created within the project and/or anything not considered to be source or final data. \cr
   \verb{data/3-final/} \tab Final data products in native R formats only. \cr
   \verb{output/} \tab Final data products for consumption outside R. \cr
   \verb{scripts/} \tab All script files. \cr
}
}
\examples{
\dontrun{
setup_project_structure()
}

}
