% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-rmd-template.R
\name{use_rmd_template}
\alias{use_rmd_template}
\title{Use an RMD template from this package}
\usage{
use_rmd_template(filename, template = "nb1", dir = "scripts")
}
\arguments{
\item{filename}{The file name for the script.}

\item{template}{The name of the RMD template (defaults to \code{"nb1"}).}

\item{dir}{The destination directory (defaults to \code{"scripts"}).}
}
\value{
\code{filename} (invisibly).
}
\description{
This is a wrapper for \code{\link[rmarkdown:draft]{rmarkdown::draft()}}. By default, it copies the
relevant files for the "nb1" RMD template into the \verb{scripts/} folder (created
by \code{\link[=setup_project_structure]{setup_project_structure()}}).
}
\examples{
\dontrun{
use_rmd_template("my-analysis")
}

}
