% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-rmd-template.R
\name{use_rmd_template}
\alias{use_rmd_template}
\title{Use an RMD template from this package}
\usage{
use_rmd_template(filename = "analysis", template = "nb1", dir = "scripts")
}
\arguments{
\item{filename}{The file name for the script.}

\item{template}{The name of the RMD template.}

\item{dir}{The destination directory.}
}
\value{
\code{filename} (invisibly).
}
\description{
This is a wrapper for \code{\link[rmarkdown:draft]{rmarkdown::draft()}}. By default, it copies the
relevant files for the "nb1" RMD template into the \verb{scripts/} folder (created
by \code{\link[=setup_project]{setup_project()}}).
}
\details{
The R Notebook template "nb1" has three related files.
\itemize{
\item \verb{_output.yml} contains the YAML options under \code{html_notebook}. These will
be shared by any markdown documents in the same directory (see \href{https://bookdown.org/yihui/rmarkdown/html-document.html#shared-options}{Shared Options, R Markdown: The Definitive Guide}).
\item \code{render.R} contains a call to \code{\link[rmarkdown:render]{rmarkdown::render()}} that's prepopulated to
execute the notebook and save the \code{.nb.html} report in the project's
\verb{output/} directory (created by \code{\link[=setup_project]{setup_project()}}).
\item \code{skeleton.Rmd} is the R Notebook file. When \code{use_rmd_template()} is
executed, the file will be saved with the name specified in the \code{filename}
argument.
}
}
\examples{
\dontrun{
use_rmd_template("my-analysis")
}

}
